Breakout PathDrawable
